<refactor>(graph) : Rename NEW/DESTROY, Add FGVAR/FGITEM
NEW and DESTROY respectively became FGNEW and FGDESTROY to better
correspond to the module they're from (FoxGraph).

FGVAR creates a graph_t/alist_t/aitem_t with
 - auto-destruction (GNU C only)
 - desired type & name
 - its corresponding constructor result assigned to it

FGITEM fetches an item's iptr from a list and casts it to desired type
to allow for better readability and less columns usage.